Refine: Modularize Rust wrapper into feature-specific modules
Splits the monolithic `lib.rs` into distinct modules for one-shot, block, streaming, and error handling APIs. This refactoring improves code organization, maintainability, and clarity by grouping related functionalities into `ctx`, `error`, `file`, `oneshot`, and `pstream` modules.

Also simplifies buffer sizing in `stdio` adapters to directly use stream-provided bounds.
